runtime2: inline error location owner type
