+++ Cargo.toml
@@ -36,7 +36,6 @@ members = [
   "crates/concurrency-caps-core",
   "crates/governance-metadata",
   "crates/linecol-core",
-  "crates/error-location-core",
   "crates/bdd-governance-core",
   "crates/bdd-governance-reporting-core",
   "crates/bdd-governance-contract",
